core bug fix and refactoring
